% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_scnas.R
\name{call_scnas}
\alias{call_scnas}
\title{Call SCNAs}
\usage{
call_scnas(rbd, config, sex)
}
\arguments{
\item{rbd, }{segment data obtained by \code{\link[=make_input]{make_input()}}.}

\item{config, }{a list of configuration parameters obtained by \code{\link[=make_input]{make_input()}}.}

\item{sex}{female or male.}
}
\value{
A list.
\itemize{
\item ploidy.adjust. a numeric vector containing upperCentral.cov, lowHDS.cov,
adj (a factor to shift lrr), ploidy and purity (the maximum prevalence).
\item est.ploidy. a numeric value. ploidy estimated from adj and purity.
\item prevalence. a tbl with cloneID and prevalence.
\item result. a tbl.
\itemize{
\item lrr50 = original lrr while lrr is adjusted one (lrr50 + log2(adj))
\item x0, y0, p0, dist0, lrr.pred0, hds.pred0. The best configuration obtained by
unconstrained values of p (config$min.prev <= p <= 1).
\item x, y, p, dist, lrr.pred, hds.pred. The final configuration derived by using
observed prevalence of clones. They are calculated only when !(x0 == 1 & y0 == 1).
}
\item deviation. sum(seg.size × dist)/100.
}
}
\description{
This function infers ploidy of the sample, tumor purity and subclones
}
