% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_xyp.R
\name{find_best_xyp}
\alias{find_best_xyp}
\title{Find the nearest copy number state}
\usage{
find_best_xyp(cfg, lrr, hds, p = NULL)
}
\arguments{
\item{cfg, }{configuration parameters}

\item{lrr}{a numeric log R ratio}

\item{hds}{a numeric Heterozygous deviation score}

\item{p}{a numeric vector. If not NULL, distance to the branches are calculated only for given p's.}
}
\value{
A tbl
}
\description{
This function finds the nearest copy number state in R-HDS plane.
Original function is findBestXYP in BubbleTree.
}
