% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clones.R
\name{find_clones}
\alias{find_clones}
\title{Estimate tumor purity and subclones}
\usage{
find_clones(rbd.adj, cfg, pladj)
}
\arguments{
\item{rbd.adj, }{segment data in which lrr has been adjusted.}

\item{cfg, }{config data obtained by \code{\link[=make_input]{make_input()}}.}

\item{pladj}{a numeric vector containing the result of \link{get_ploidy_adjustment}.}
}
\value{
A list.
}
\description{
This function estimates tumor purity and subclones as well as the best copy
number states. Original function is findClones in BubbleTree.
}
