% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_coord.R
\name{get_genomic_coord}
\alias{get_genomic_coord}
\title{Plot allele-specific copy number changes across genome}
\usage{
get_genomic_coord(df, hg38.seqinfo, space.skip = 0.1)
}
\arguments{
\item{df}{A tbl.}

\item{hg38.seqinfo}{Use \code{data("centromere.hg38")}}

\item{space.skip}{Space between chromosomes (default: 0.01)}
}
\value{
A ggplot object
}
\description{
Plot allele-specific copy number changes across genome
}
