% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ploidy_adjustment.R
\name{get_ploidy_adjustment}
\alias{get_ploidy_adjustment}
\title{Infer sample ploidy and estimate adjustment factor for copy ratio}
\usage{
get_ploidy_adjustment(rbd, cfg)
}
\arguments{
\item{rbd, }{segment data obtained by \code{\link[=make_input]{make_input()}}.}

\item{cfg, }{config data obtained by \code{\link[=make_input]{make_input()}}.}
}
\value{
A named vector.
}
\description{
This function infers sample ploidy and estimate adjustment factor for copy ratio.
Original function is getPloidyAdj in BubbleTree.
}
