% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scna_states.R
\name{get_scna_states}
\alias{get_scna_states}
\title{Estimate tumor purity and subclones}
\usage{
get_scna_states(rbd.adj, cfg, p = NULL)
}
\arguments{
\item{rbd.adj, }{segment data obtained by \link{get_ploidy_adjustment}, in which lrr has been adjusted.}

\item{cfg, }{config data obtained by \code{\link[=make_input]{make_input()}}.}

\item{p}{a numeric vector. If not NULL, distance to the branches are calculated only for given p's.}
}
\value{
A list.
}
\description{
This function estimates the most likely configuration of copy number states
}
