% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xypGrid.R
\name{get_xypGrid}
\alias{get_xypGrid}
\title{Get a grid of copy number states: (x, y, p)}
\usage{
get_xypGrid(max.ploidy, min.prev)
}
\arguments{
\item{max.ploidy}{maximum ploidy to consider.}

\item{min.prev}{minimum prevalence.}
}
\value{
A tbl
}
\description{
Get a grid of copy number states: (x, y, p)
}
