% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input.R
\name{make_input}
\alias{make_input}
\title{Make input data}
\usage{
make_input(input_dir, tumorID, sex)
}
\arguments{
\item{input_dir}{Input directory where \if{html}{\out{<tumorID>}}.modelFinal.seg, \if{html}{\out{<tumorID>}}.hets.tsv and
\if{html}{\out{<tumorID>}}.denoisedCR.tsv can be found.}

\item{tumorID}{Sample ID of tumor.}

\item{sex}{female, male or a filename that contains a word, female or male.
If missing, sex is assumed to be male and only autosomes are analyzed.}
}
\value{
A list containing:
\itemize{
\item tumorID.
\item sex.
\item rbd, input for the SCNA calling.
\item cnv.nohets, segments without heterozygous sites.
\item crs, copy ratio data used for plotting.
\item segs, segmentation result from GATK where seg.id has been added and
posterior MAFs have been converted to HDS.
\item mafs, MINOR_ALLELE_FRACTION_POSTERIOR_50 for plotting.
\item hets, ALT allele frequency at heterozygous sites in normal sample.
}
}
\description{
This function generates dataset to be used in this package.
}
