% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ASCN.R
\name{plot_ASCN}
\alias{plot_ASCN}
\title{Plot allele-specific copy number changes across genome}
\usage{
plot_ASCN(
  res,
  centromere.hg38,
  hg38.seqinfo,
  min.prev = 0.2,
  ymax = 4.5,
  baseline.adj = 0.05,
  space.skip = 0.005
)
}
\arguments{
\item{res}{result of \link{call_scnas}.}

\item{centromere.hg38}{Use \code{data("centromere.hg38")}.}

\item{hg38.seqinfo}{Use \code{data("centromere.hg38")}}

\item{min.prev}{minimum prevalence.}

\item{ymax}{Maximum copy number of specific allele (default: 4.3).}

\item{baseline.adj}{Upward (y = n(B)) and downward (x = n(A)) shift of A and B allele (default: 0.1).}

\item{space.skip}{Space between chromosomes (default: 0.01)}
}
\value{
A ggplot object
}
\description{
Plot allele-specific copy number changes across genome
}
