% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_copyRatio.R
\name{plot_copyRatio}
\alias{plot_copyRatio}
\title{Plot copy ratio data}
\usage{
plot_copyRatio(
  res,
  centromere.hg38,
  hg38.seqinfo,
  crs = NULL,
  segs = NULL,
  min.prev = 0.2,
  ymax = 3,
  baseline.adj = 0.1,
  space.skip = 0.005
)
}
\arguments{
\item{res}{result of \link{call_scnas}.}

\item{centromere.hg38}{Use \code{data("centromere.hg38")}.}

\item{hg38.seqinfo}{Use \code{data("centromere.hg38")}.}

\item{crs}{denoised copy ratio data (seqnames, start, end, lrr).}

\item{segs}{segment data (seqnames, start, end, lrr) that may include segments filtered by btpredict}

\item{min.prev}{minimum prevalence (default: 0.2).}

\item{ymax}{Maximum copy ratio (default: 3).}

\item{baseline.adj}{Upward (y = n(B)) and downward (x = n(A)) shift of A and B allele (default: 0.1).}

\item{space.skip}{Space between chromosomes (default: 0.01)}
}
\value{
A ggplot object
}
\description{
Plot copy ratio data
}
