% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HetLOH.R
\name{plot_HetLOH}
\alias{plot_HetLOH}
\title{Plot LOH data}
\usage{
plot_HetLOH(
  res,
  centromere.hg38,
  hg38.seqinfo,
  hets = NULL,
  min.prev = 0.2,
  baseline.adj = 0.1,
  space.skip = 0.005
)
}
\arguments{
\item{res}{result of \link{call_scnas}.}

\item{centromere.hg38}{Use \code{data("centromere.hg38")}.}

\item{hg38.seqinfo}{Use \code{data("centromere.hg38")}.}

\item{hets}{A tbl containing heterozygous site data (seqnames, start, end, freq).}

\item{min.prev}{minimum prevalence (default: 0.2).}

\item{baseline.adj}{Upward (y = n(B)) and downward (x = n(A)) shift of A and B allele (default: 0.1).}

\item{space.skip}{Space between chromosomes (default: 0.01)}
}
\value{
A ggplot object
}
\description{
Plot LOH data
}
