% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RvsHDS.R
\name{plot_RvsHDS}
\alias{plot_RvsHDS}
\title{Plot copy ratio vs HDR}
\usage{
plot_RvsHDS(res, tumorID, min.prev = 0.2, max.size = 20)
}
\arguments{
\item{res}{result of \link{call_scnas}.}

\item{tumorID}{Tumor ID.}

\item{min.prev}{minimum prevalence (default: 0.2).}
}
\value{
A ggplot object
}
\description{
Plot copy ratio vs HDR
}
