% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xypGrid.R
\name{plot_xypGrid}
\alias{plot_xypGrid}
\title{Plot a grid of copy number states: (x, y, p)}
\usage{
plot_xypGrid(max.ploidy, min.prev)
}
\arguments{
\item{max.ploidy}{maximum ploidy to consider.}

\item{min.prev}{minimum prevalence.}
}
\value{
A ggplot object
}
\description{
Plot a grid of copy number states: (x, y, p)
}
